feat(pid): migrate PIDs from u32 to u64 time-seeded values for reboot persistence
Replaces sequential u32 PIDs (which reset to 1 on every kernel restart, causing
session/invocation collisions) with time-seeded u64 PIDs generated via
`Pid::generate()`: 42-bit ms timestamp since 2025-01-01 | 22-bit random salt.

Key changes:
- `Pid(u32)` → `Pid(u64)` with `Pid::generate()`, `Pid::from_u64()`, `Pid::kernel()`
- `allocate_pid()` removed; spawn uses `Pid::generate().as_u64()` inline
- `SessionRecord.owner_pid/pids`, `InvocationRecord.pid`, `IssuedTo.pid`,
  `AgentRecord.pid`, `CapturedBy::Agent`, `MemoryRecordMetadata.agent_pid`,
  `CompleteMetadata.agent_pid` all widened to u64
- All internal HashMaps keyed on pid changed from `HashMap<u32, _>` to `HashMap<u64, _>`
- ATP wire format: `ATPCommand::AgentKill/AgentStatus.pid` → `String` (D3 decision:
  avoids JS float precision loss for u64 > 2^53); event bus emits `"pid": pid.to_string()`
- Architecture docs updated: 06-agents.md, 14-agent-persistence.md

All 1636 avix-core tests pass.

+++ crates/avix-core/src/executor/memory.rs
@@ -136,7 +136,7 @@ impl MemoryManager {
     /// Write a session summary to episodic memory when SIGSTOP fires.
     pub async fn auto_log_session_end(
         &self,
-        pid: u32,
+        pid: u64,
         agent_name: &str,
         spawned_by: &str,
         session_id: &str,
